mod to XYPatternBlock
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@2046 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_geometry/XYPatternBlock.cpp
@@ -103,7 +103,7 @@ int XYPatternBlock::setLanePoints()
   double alter_ang2 = angle360(m_angle + 90);
   
   double lane_pts_count_d = (m_blockwid / m_swathwid) + 1; 
-  int i, lane_pts_count = trunc(lane_pts_count_d + 0.99999);
+  int i, lane_pts_count = (int)(trunc(lane_pts_count_d + 0.99999));
 
   double cptx, cpty;
   idPointCenter(cptx, cpty);
